conda_requirements.txt
beautifulsoup4==4.4.1
geopy==1.11.0
grequests==0.3.0
humanize==0.5.1
networkx==1.11
nltk==3.2.1
pandas>=0.18.1
scikit-learn==0.18.1
seaborn==0.7.1
ipython==5.1.0
